Convert AnnData to SingleCellExperiemnt using scDIOR/zellkonverter/schard.

Convert AnnData to SingleCellExperiemnt using scDIOR/zellkonverter/schard.

Usage

AD2SCE(
  anndata.file,
  method = c("scDIOR", "zellkonverter", "schard"),
  assay = "RNA",
  slot = "counts",
  use.raw = TRUE,
  conda.path = NULL
)

Arguments

anndata.file

The file contains AnnData.

method

Method used to perform conversion, choose from "scDIOR", "zellkonverter", "schard". Default: "scDIOR".

assay

The type of data, used when method is "scDIOR". Default: "RNA" (scRNA-seq data).

slot

Name used when saving adata.X as an assay, used when method is "zellkonverter". Default: "counts".

use.raw

Logical value, whether to use adata.raw. Default: TRUE. When method is "scDIOR" and use.raw is TRUE (FALSE), raw.X -> assays (X/layers -> assays); When method is "zellkonverter" and use.raw is TRUE (FALSE), raw.X -> altExp, X and layers -> assays (X and layers -> assays); When method is "schard" and use.raw is TRUE (FALSE), raw.X -> assays (X -> assays).

conda.path

Conda environment path, used when method is "scDIOR" or "zellkonverter". Default: NULL.

Value

A SingleCellExperiment object.

Examples

# scDIOR
sce.scdior <- AD2SCE(anndata.file = "pbmc3k.h5ad", method = "scDIOR", assay = "RNA", use.raw = TRUE, conda.path = "/path/to/conda")
# zellkonverter
sce.zell <- AD2SCE(anndata.file = "pbmc3k.h5ad", method = "zellkonverter", slot = "scale.data", use.raw = TRUE, conda.path = "/path/to/conda")
# schard
sce.schard <- AD2SCE(anndata.file = "pbmc3k.h5ad", method = "schard", use.raw = TRUE)