Extract Raw Count Matrix or Fortmat Supplementary Files to 10x.

Extract Raw Count Matrix or Fortmat Supplementary Files to 10x.

Usage

ExtractGEOExp(
  pf.obj,
  acce,
  supp.idx = 1,
  down.supp = FALSE,
  timeout = 3600,
  supp.type = c("count", "10x", "10xSingle"),
  file.regex = NULL,
  extra.cols = c("chr", "start", "end", "strand", "length", "width", "chromosome",
    "seqnames", "seqname", "chrom", "chromosome_name", "seqid", "stop"),
  transpose = TRUE,
  out.folder = NULL,
  accept.fmt = c("MEX", "h5"),
  gene2feature = TRUE
)

Arguments

pf.obj

GEO object of platform.

acce

GEO accession number.

supp.idx

The index of supplementary files to download. Default: 1.

down.supp

Logical value, whether to download supplementary files to create count matrix. If TRUE, always download supplementary files. If FALSE, use ExpressionSet (If contains non-integer or emoty, download supplementary files automatically). Default: FALSE.

timeout

Timeout for download.file. Default: 3600.

supp.type

The type of downloaded supplementary files, choose from count (count matrix file or single count matrix file), 10x (cellranger output files in tar/gz supplementary files, contains barcodes, genes/features and matrix, e.g. GSE200257) and 10xSingle (cellranger output files in supplementary files directly, e.g. GSE236082). Default: count.

file.regex

The regex to extract correct count matrix files. Default: NULL.

extra.cols

Extra columns to remove, e.g., "Chr", "Start", "End", "Strand", "Length" (featureCounts). Default: "chr", "start", "end", "strand", "length", "width", "chromosome", "seqnames", "seqname", "chrom", "chromosome_name", "seqid", "stop".

transpose

Logical value, whether to transpose the matrix. Used when the number of rows is less than the number of columns. Default: TRUE.

out.folder

Output folder to save 10x files. Default: NULL (current working directory).

accept.fmt

Vector of accepted 10x output format, MEX (barcode/feature/gene/matrix), h5 (h5/h5.gz). Default: c("MEX", "h5").

gene2feature

Logical value, whether to rename genes.tsv.gz to features.tsv.gz. Default: TRUE.

Value

Count matrix (supp.type is count) or NULL (supp.type is 10x/10xSingle).