Extract Metadata of scRNA-seq Datasets in PanglaoDB.

Extract Metadata of scRNA-seq Datasets in PanglaoDB.

Usage

ExtractPanglaoDBMeta(
  species = NULL,
  protocol = NULL,
  tissue = NULL,
  cell.num = NULL,
  show.cell.type = TRUE,
  local.data = TRUE
)

Arguments

species

The species of the datasets, choose from "Homo sapiens", "Mus musculus", one or multiple value. Default: NULL (All).

protocol

Protocol used to generate the datasets, choose from "10x chromium", "drop-seq", "microwell-seq", "C1 Fluidigm", "inDrops", "Smart-seq2", "CEL-seq", one or multiple value. Default: NULL (All).

tissue

The tissue of the datasets, obtain available values with StatDBAttribute. Default: NULL (All).

cell.num

Cell number filter. If NULL, no filter; if one value, lower filter; if two values, low and high filter. Deault: NULL.

show.cell.type

Logical value, whether to show inferred cell type. Default: TRUE.

local.data

Logical value, whether to use local data (PanglaoDB is no longer maintained). Default: TRUE.

Value

Dataframe contains SRA, SRS, Tissue, Protocol, Species, Cells, CellType (inferred).

Examples

# \donttest{
human.meta <- ExtractPanglaoDBMeta(
  species = "Homo sapiens",
  protocol = c("Smart-seq2", "10x chromium"),
  cell.num = c(1000, 2000)
)
# }