Load Objects, Count Matrix, and Metadata from RData.

Load Objects, Count Matrix, and Metadata from RData.

Usage

LoadRData(
  rdata,
  accept.fmt = c("Seurat", "seurat", "SingleCellExperiment", "cell_data_set",
    "CellDataSet", "DESeqDataSet", "DGEList"),
  show.object = TRUE,
  return.obj = TRUE,
  slot = c("counts", "data", "scale.data")
)

Arguments

rdata

Path to RData file.

accept.fmt

Vector, the format of objects for loading. Default: c("Seurat", "seurat", "SingleCellExperiment", "cell_data_set", "CellDataSet", "DESeqDataSet", "DGEList"). "Seurat" for Seurat v3, v4; "seurat" for Seurat v2.

show.object

Logical value, whether to show the class of available objects. Default: TRUE.

return.obj

Logical value, whether to load the available objects in accept.fmt to global environment. Default: TRUE.

slot

Vector, the type of count matrix to pull. Default: c("counts", "data", "scale.data"). 'counts': raw, un-normalized counts, 'data': normalized data, scale.data: z-scored/variance-stabilized data.

Value

List contains count matrix and metadta or NULL.

Examples

if (FALSE) { # \dontrun{
# download RData from GEO
ParseGEOProcessed(acce = "GSE244572", timeout = 360000, supp.idx = 1, out.folder = "/path/to/RData", file.ext = c("rdata", "rds", "h5ad", "loom"))
# with six assays: "RNA", "ADT", "nADT", "SCT", "integrated", "IADT"
multi.assyas.list <- LoadRData(
  rdata = "/path/to/GSE244572_RPE_CITESeq.RData",
  accept.fmt = c("Seurat", "seurat", "SingleCellExperiment", "cell_data_set", "CellDataSet", "DESeqDataSet", "DGEList"),
  slot = "counts"
)
# only one object in GSE244572_RPE_CITESeq.RData
names(multi.assyas.list)
# metadata
multi.assyas.list$obj$meta.data %>% head()
# assay names
names(multi.assyas.list$obj$count.mat)
# raw count matrix (slot)
multi.assyas.list$obj$count.mat$RNA$counts[1:5, 1:5]
} # }