Download Processed Objects from GEO.

Download Processed Objects from GEO.

Usage

ParseGEOProcessed(
  acce,
  timeout = 3600,
  supp.idx = 1,
  file.ext = c("rdata", "rds", "h5ad", "loom"),
  out.folder = NULL,
  return.seu = FALSE,
  merge = TRUE
)

Arguments

acce

GEO accession number.

timeout

Timeout for download.file. Default: 3600.

supp.idx

The index of supplementary files to download. Default: 1.

file.ext

The valid file extension for download (ignore case and gz suffix). When NULL, use all files. Default: c("rdata", "rds", "h5ad").

out.folder

The output folder. Default: NULL (acce folder under current working directory).

return.seu

Logical value, whether to load downloaded datasets to Seurat. Valid when rds in file.ext and all datasets download successfully. Default: FALSE.

merge

Logical value, whether to merge Seurat list when there are multiple rds files, used when return.seu is TRUE. Default: FALSE.

Value

SeuratObject (return.seu is TRUE, rds in file.ext) or NULL (return.seu is FALSE or rds not in file.ext).

Examples

if (FALSE) { # \dontrun{
# need users to provide the output folder
# suitable for rdata, rdata.gz, rds, rds.gz
# return SeuratObject when return.seu = TRUE
GSE285723.seu <- ParseGEOProcessed(
  acce = "GSE285723", supp.idx = 1,
  file.ext = c("rdata", "rds"), return.seu = TRUE,
  out.folder = "/path/to/outfoder"
)
geo.rdata.log <- ParseGEOProcessed(
  acce = "GSE311825", supp.idx = 4,
  file.ext = c("rdata", "rds"),
  out.folder = "/path/to/outfoder"
)
# suitable for h5ad, h5ad.gz
geo.h5ad.log <- ParseGEOProcessed(
  acce = "GSE311813", supp.idx = 1,
  file.ext = c("h5ad"),
  out.folder = "/path/to/outfoder"
)
# suitable for loom, loom.gz
geo.loom.log <- ParseGEOProcessed(
  acce = "GSE286325", supp.idx = 1,
  file.ext = c("loom"),
  out.folder = "/path/to/outfoder"
)
} # }