Run CellRanger on Downloaded FASTQ Files.

Run CellRanger on Downloaded FASTQ Files.

Usage

RunCellRanger(
  sample.dir,
  ref,
  localcores = 4,
  localmem = 16,
  out.folder = NULL,
  cr.path = NULL,
  cr.paras = "--chemistry=auto --jobmode=local"
)

Arguments

sample.dir

Directory contains all samples.

ref

Path of folder containing 10x-compatible transcriptome reference (--transcriptome).

localcores

Set max cores the pipeline may request at one time (--localcores). Only applies to local jobs (--jobmode=local). Default: 4.

localmem

Set max GB the pipeline may request at one time (--localmem). Only applies to local jobs (--jobmode=local). Default: 16.

out.folder

Output folder. Default: NULL (current working directory).

cr.path

Path to cellranger. Default: NULL (conduct automatic detection).

cr.paras

Parameters for cellranger. Default: "–chemistry=auto –jobmode=local".

Value

Vector contains failed samples or NULL.

Examples

if (FALSE) { # \dontrun{
RunCellRanger(
  sample.dir = "/path/to/fastq", # GSMXXXX
  ref = "/path/to/cellranger_tiny_ref/3.0.0",
  out.folder = "/path/to/results",
  cr.path = "/path/to/cellranger-x.x.x/cellranger"
)
} # }